BUG: close shelve file using a context manager
Using a context manager ensures:

- that the shelve file will be closed, and
- that the directory tree that contains the
  shelve file will be deleted after the file
  has been closed.


## The bugs

The issue is illustrated with the following script,
which opens a file and raises an exception before closing the
file. The finalization section deletes the directory that
contained the file.

```python
"""Test closing of file.

To observe the `ResourceWarning`,
run this script with `python -X dev`.
"""
import os
import shutil


def main():
    shelve_dir = 'temp_dir'
    fname = os.path.join(shelve_dir, 'temp_file')
    os.makedirs(shelve_dir)
    try:
        cache = open(fname, 'w')
        raise AssertionError()
        cache.close()
    finally:
        shutil.rmtree(shelve_dir)


if __name__ == '__main__':
    main()
```

This behavior is erroneous in two ways:

1. the file is not closed
2. the directory is deleted while
   a file in that directory is still open.

[On certain operating systems][1], unlinking a file does not
delete the file before all programs that have opened it
close it. Unlinking removes the file name from the
directory listing, but deletion can happen later.
Deletion can be postponed when the file is still open.

I assume this is the reason that I observe an
`AssertionError` and a `ResourceWarning` about an unclosed
file, but no error from `shutil.rmtree`,
even though `shutil.rmtree` deletes a directory that
contains a file that is still open.
Perhaps on other operating systems
`shutil.rmtree` would raise an error.

The [documentation of the function `shutil.rmtree`][2]
does not specify how it works, so I assume that
`shutil.rmtree` calls `unlink`.


## Corrected version

The following script does not leave the file open,
and so does not issue a `ResourceWarning`.

```python
"""Test closing of file."""
import os
import shutil

def main():
    shelve_dir = 'temp_dir'
    fname = os.path.join(shelve_dir, 'temp_file')
    os.makedirs(shelve_dir)
    try:
        with open(fname, 'w') as cache:
            raise AssertionError()
    finally:
        shutil.rmtree(shelve_dir)

if __name__ == '__main__':
    main()
```

Also, on other operating systems, the above script
avoids the deletion of an open file by `shutil.rmtree`,
and so any such errors.


## Unclosed shelve files

In the examples above, I used the function `open`,
whereas the error that this commit corrects
was due to a file opened using the function `shelve.open`.

I did not observe any `ResourceWarning` due to an
unclosed *shelve* file (opened by calling `shelve.open`).
The reason may be Python automatically closing the shelve file.

In any case, as noted in the documentation of the function
[`shutil.open`][3], open shelve files should be explicitly
closed. Quoting from the documentation:

> **Note:** Do not rely on the shelf being closed automatically;
> always call `close()` explicitly when you don’t need it any more,
> or use `shelve.open()` as a context manager:

Previously the database file for the shelf was explicitly closed.
However, the call to the method `shelve.Shelf.close` was at the
wrong place, because any exception raised before
reaching that `close` call would have left
the shelve file open. This is why it is necessary
to use a context manager in a `with` statement.

As the [documentation of the `with` statement][4] notes,
using `with` ensures that the context manager's `__exit__`
method will be called, even if the suite of statements
has raised an exception.

Also, from the wording of the documentation quoted above
("Do not rely on the shelf being closed automatically")
it seems that Python may in some cases automatically close
shelve files. This may explain why a `ResourceWarning`
was not issued in my attempts to mimick the above scripts
for a shelve file.


## Multiline `with` suite

I used a backslash (`\`) to write what is
between the keyword `with` and the colon that
starts the [suite of statements][4] of the `with`
statement.

In Python 3.9, parentheses can be used,
instead of a backslash:

```python
        with (shelve.open(tmp_fname) as cache,
                open(file_name, 'w') as f):
```

but this syntax does not work on earlier Python versions.

I used a backslash, instead of parentheses,
in order to support Python < 3.9.


## Shelves as context managers

This change is incompatible with Python 2.7,
because context manager support was added to the
[class `shelve.Shelf`][5] in Python 3.4 ([read also][6]).

The alternative would have been to write
a nested `with`-statement. Support for Python 2.7
will soon be removed, so there is little reason
to introduce complexity for maintaining compatibility.

So dumping and loading BDDs to JSON files
will now be supported only on Python 3.

Using a context manager to close the shelve file,
*and* remaining compatible with Python 2.7 is possible.
By subclassing the class `shelve.Shelf` and adding
context-manager methods.
But it is simpler to require Python 3.

Note in the README that a few parts require Python 3.

Skip in Python 2.7 tests that involve dumping or
loading of BDDs to or from JSON files.

I assume that the purpose of not using shelf objects
as context managers in commit
    61b60bfedccf45de418641f53caec204608dd714
(which introduced this behavior) was to
remain compatible with Python 2.7.

Even if that was the case, still this was a bug
that could have been avoided by subclassing
the class `shelve.Shelf`, as described above.


[1]: https://en.wikipedia.org/wiki/Unlink_(Unix)
[2]: https://docs.python.org/3/library/shutil.html#shutil.rmtree
[3]: https://docs.python.org/3/library/shelve.html#shelve.open
[4]: https://docs.python.org/3/reference/compound_stmts.html#with
[5]: https://docs.python.org/3/library/shelve.html#shelve.Shelf
[6]: https://docs.python.org/2.7/library/shelve.html#shelve.open