BUG: rm assertion that prevents execution of prototype method
The method `dd.cudd_zdd.ZDD._quantify_optimized` had an
assertion that allowed only calls for existential quantification.
It is unclear why this assertion was added. Perhaps it was
added at an early stage of development. However, this function
currently appears to support universal quantification.

In any case, this is a prototype implementation. Users of the
module `dd.cudd_zdd` that call quantification invoke another
function. So this change does not relate to user code
(since the identifier `_quantify_optimized` is hidden [PEP 8][pep8]).

[pep8]: https://www.python.org/dev/peps/pep-0008/