BUG: initialize pointer first, to be prepared for `__dealloc__`
Previously, an exception raised inside the method `__cinit__`
resulted in the attribute `manager` being an uninitialized pointer.

The method `__dealloc__` is always executed, including also
when an exception is raised when Python attempts to execute
the method `__cinit__`. Thus, the execution of `__dealloc__`
in this case resulted in a "Segmentation fault: 11", because
`__dealloc__` attempted to call the C function `Cudd_CheckZeroRef`
with an uninitialized pointer as argument.

(The method `__dealloc__` also calls the C function `Cudd_Quit`.
According to the docstring of the function `Cudd_Quit`,
calling `Cudd_Quit` with a `NULL` pointer has no effect.
Therefore, a segmentation fault is raised only due to calling
the function `Cudd_CheckZeroRef`. The docstring is in the
file `cudd-3.0.0/cudd/cuddInit.c`.)

By initializing the attribute `manager` to `NULL` before doing
anything else inside the method `__cinit__`, the segmentation
fault is avoided inside `__dealloc__` by checking whether the
attribute value is `NULL`, and a `RuntimeError` is raised instead.

Interestingly, it is possible for an exception to be raised
even before executing the method `__cinit__`. For example by
writing `dd.cudd.BDD(1, 2, 3)`. In this case, the method
`__dealloc__` is still called. The pointer attribute `manager`
would in this case be uninitialized (an indeterminate value),
thus the check for the value `NULL` is not guaranteed to
avoid the segmentation fault.

To avoid this case, the signature of the method `__cinit__`
is extended with `*arg, **kw`. This ensures that the above
initialization `dd.cudd.BDD(1, 2, 3)` does not prevent the
method `__cinit__` from starting to execute, so that the
attribute `manager` be initialized to `NULL`.

The behavior of Cython described above has been discussed in the
thread ["is __cinit__ guaranteed to complete?"](
    https://groups.google.com/g/cython-users/c/TXrle75lE6Q/m/xmVi4oZa5doJ)
on https://groups.google.com/g/cython-users.

That the method `__dealloc__` is always called is implied in
the [Cython tutorial](
    http://conference.scipy.org/proceedings/scipy2009/paper_1/full_text.pdf)
from the SciPy Conference of 2009, specifically page 10.

[Documentation of the method `__dealloc__`](
    https://docs.cython.org/en/latest/src/userguide/special_methods.html#finalization-method-dealloc).