BUG: pass `roots` to `to_nx` that are not random
Previously, `roots=[u]`, where `u` was whatever value
resulted from the set comprehension a few lines above
the line that calls the function `to_nx`.

This bug was introduced in commit:
    917f870fee1c7059421ee7839d687021eb454762

It is unclear what my exact intention for plotting
at that commit was. The function `bdd_to_mdd` operates on
an entire BDD manager, i.e., no root nodes are given
as arguments to `bdd_to_mdd`. So there are no roots to
give to the function `to_nx`.

Moreover, within the function `bdd_to_mdd`,
the variable `u` is used only as a [loop variable][1],
in comprehensions and iterations. Therefore, the
variable `u` cannot be used to define the argument `roots`
when calling the function `to_nx`.

This commit uses all nodes in the BDD manager `bdd`,
because it seems that the intention was to color
a subset of the BDD nodes, and dump the resulting graph.

Note that `roots=bdd` passes a container to the
function `to_nx`, and the function `to_nx` iterates
over `bdd`.

[1]: https://docs.python.org/3/glossary.html#term-generator-expression