API: extract unused code from function `bdd_to_mdd`
and rewrite this code as the new function `dd.mdd._debug_dump`.

As a consequence, `networkx` is now not required to
call the function `dd.mdd.bdd_to_mdd`. Previously,
`networkx` was used when calling the function `bdd_to_mdd`,
but the resulting `networkx` graph was not used.

Commit 917f870fee1c7059421ee7839d687021eb454762 introduced
the code that is now rewritten as part of the function
`_debug_dump`. When this code was introduced, the
resulting `networkx` graph *was* actually used.

In more detail, the `networkx` graph was converted to
a `pydot` graph, and the `pydot` graph written to a PDF,
using GraphViz. This was in commit
917f870fee1c7059421ee7839d687021eb454762.

This is why the newly added function `_debug_dump` includes
an updated version of code from commit
917f870fee1c7059421ee7839d687021eb454762, so that the
`networkx` graph is now converted to a `pydot` graph,
and written to a PDF using GraphViz.

The code, from commit 917f870fee1c7059421ee7839d687021eb454762,
that converted from `networkx` to `pydot`,
and then wrote to a PDF, was commented in commit
3bb95afdc98faf73eaf9d03a7ce5c6b67837c9af
and removed (together with the then unused `networkx` import)
in commit 5f2543b496baff42d22ecb8e076b19151315e6ff.

An import of `networkx` now happens inline, inside the
function `_debug_dump`.

The rewritten code also includes additional assertions,
and improved names of variables.

The newly added function `_debug_dump` is not called from
anywhere in the package `dd`. It exists only for debugging
purposes. A commented call to `_debug_dump` is added
within the function `bdd_to_mdd`, as a reminder of where
`_debug_dump` is intended to be called from, during debugging.

This change is marked with "API" due to the change in
requirements for calling the function `dd.mdd.bdd_to_mdd`.
This API change has no direct effect on users.