API: reduce the memory needed by `dd.cudd_zdd.to_nx`
This function was storing node indices in both:
- a `set` named `visited`, and
- as keys of a `dict` named `umap`.

Doing so was redundant.

The important detail for using only the dictionary `umap`,
and omitting the set `visited`, is to recurse *before*
adding edges from node `u` to its successor nodes
(nodes `v` and `w`).

Note that a side-effect of this change is that
the indexing of nodes in the resulting DOT graph
is changed. The node labels though remain the same.

This indexing change is not visible in the PDF files
written by the method `dd.cudd_zdd.ZDD.dump`.
This method calls the function `dd.cudd_zdd.to_nx`
(and dumping DOT source text is not currently supported).

However, the indexing of nodes in the graph returned
by the function `dd.cudd_zdd.to_nx` did change by
this commit, which is why this commit is marked
with "API".

Nevertheless, the indexing of nodes in the graph
returned by the function `dd.cudd_zdd.to_nx`
did not have any particular meaning as part of
the API. Therefore, this change is not
expected to have an effect on users
other than that the resulting graph will have the
indices permuted over the graph's nodes
(i.e., the result is isomorphic to before).

Other improvements in this rewriting of the functions
`dd.cudd_zdd.to_nx` and `dd.cudd_zdd._to_nx`:

- improve variable names

- index nodes using `len(umap)`, instead of `len(g)`

- return earlier from within the function `_to_nx`:
  no need to get the successor nodes
  if the node has already been visited

- recognize leaf nodes by testing for `u.var is None`,
  because this is more readable than `v is None`,
  where `v` a successor node

- skip getting the successors for leaf nodes

- call `dict.setdefault` only once:
  the other times use the method `dict.__getitem__`,
  which thus serves also as an assertion.

  Moreover, this limits the index creation via `len(umap)`
  at a single line of code, instead of 3 lines of code
  (which would otherwise have to remain in sync
  in case the method of index creation changes).

- use a conditional (`if` statement) to select the node's
  label, and then call the method
  `networkx.MultiDiGraph.add_node` only once.

  Previously, the method `add_node` was called at two
  different lines within the function `_to_nx`,
  which were 6 lines apart, and thus the implemented
  algorithm was less evident.

- do not store `umap` as a variable within the
  function `to_nx`: this allows the `dict` to
  be garbage-collected by Python as soon as the
  function `_to_nx` returns.

  Although this does not make a difference the way
  that the function `to_nx` is currently written
  (since the scope of `to_nx` is exited immediately
  after `_to_nx` returns), nonetheless it is good
  practice to do this.