BUG: correct unreachable code in `cudd_zdd._disjoin()`
and in `cudd_zdd._conjoin()`. Changes are similar,
so `_disjoin()` is described below. Similar also to commit:

> BUG: correct unreachable code in `cudd_zdd.ZDD._disjoin()`

An `AssertionError` is raised, instead of a `ValueError`,
because the recursion of `_disjoin()` starts by
`_disjoin_root()`, which correctly passes `level=0` as
argument. So the error implies an error in the traversal,
not an error in arguments of `_disjoin_root()`.

The new assertions implement the intended API of
`_disjoin()` and `_conjoin()`. The API is unchanged.

New functions for calling `cdef` functions from Python:
- `dd.cudd_zdd._call_disjoin()`
- `dd.cudd_zdd._call_conjoin()`

These functions are intended for only testing purposes.

The error was introduced in commit
`2fb70ab91d14ef1ac0722a32a0a04f91acaad73f`,
which introduced the module `dd.cudd_zdd`.


## Unreachable code

In `_disjoin()`, the `if` statement:

```
if v == DD_ONE(mgr) and index == -1:
```

was unreachable via `_disjoin_root()`,
because then the following predicate was true:

```
(u != DD_ZERO(mgr) and
 v != DD_ZERO(mgr) and
 (u != DD_ONE(mgr) or index != -1))
```

Thus:

- CASE `index != -1`
  `~ (v == DD_ONE(mgr) and index == -1)`
  which is the `if` predicate.

- CASE `index == -1`
  `u != DD_ONE(mgr)`

  For a call to `_disjoin()` via `_disjoin_root()`,
  `u == DD_ZERO(mgr) or u == DD_ONE(mgr)` when
  `index == -1 or index == CUDD_CONST_INDEX`.

  By the predicate above `u != DD_ZERO(mgr)`,
  so `u == DD_ONE(mgr) and index == -1`.
  Thus the `if` statement is not reached.

No direct calls to `_disjoin()` are in `dd.cudd_zdd`.
Therefore, lines inside this `if` were unreachable.