BUG: `print` is a function in Python 3
Introduced in commit 2fb70ab91d14ef1ac0722a32a0a04f91acaad73f.
This is a syntax error in Python 3.

This typo did not cause issues with Cython up to now,
because in the module `download.py`,
the function `extensions` passes `language_level=2`
in the argument `compiler_directives` to the
function `Cython.Build.cythonize`.