BUG: unsupported cases of `other is None` in `__richcmp__()`
Before this commit, when `other=None` was passed to:
- `dd.cudd.Function.__richcmp__()` and
- `dd.cudd_zdd.Function.__richcmp__()`
they did not raise the correct exceptions.

This commit corrects `__richcmp__()` so that now it
detects these cases of erroneous input, and raises a
`ValueError` with a more informative message than
the messages of the exceptions that were raised
before this commit.


## Description of error

`dd.cudd.Function.__richcmp__()` raised either a
suitable type of exception from an unexpected location,
and not with the intended message:

```python
u < None
    # ValueError: `self.manager != other.manager`
u <= None
    # ValueError: `self.manager != other.manager`
```

or another type of exception:

```python
u > None
    # TypeError: bad operand type for unary ~: 'NoneType'
u >= None
    # TypeError: bad operand type for unary ~: 'NoneType'
```

This bug was introduced in commit
`a858fa756cb0b27d74cf113ff73efadc05c8b856`.