BUG: handle unhashable values when asserting
The original motivation for the changes in
this commit was to avoid instantiating
a temporary `set`, and iterate instead.

Given the current implementation of the
function `dd.cudd._cube_array_to_dict()`,
a `TypeError` due to unhashability of
`support.values()` is impossible
(they are strings that are BDD variable names).

Nonetheless, this approach also avoids
instantiating a temporary `set`.