API: return `list` from `BDD.load()` of Pickle file
This change affects the signatures of the methods:

- `dd.bdd.BDD.load()`
- `dd.autoref.BDD.load()`

The earlier return value from `dd.bdd.BDD.load()`
was intended to provide the mapping from
dumped node references to loaded node references.

With this mapping, it was possible to
map information that existed when the
BDD was dumped, and that has been stored
separately in another file, and loaded again
in this run of Python, to nodes that exist
in the BDD manager in this run of Python.

This mapping is still possible now,
by using a `list` of roots. Only the
roots of interest need be included in
this `list` (instead of returning
a `dict`--previously called `umap`
within the method
`dd.bdd.BDD._load_pickle()`--with
all the nodes that are reachable from
the roots that were given to the
method `dd.bdd.BDD.dump()`).

In case all nodes reachable from the
roots are needed in the returned `list`,
then they can always be passed as roots
at the time of dumping the BDD.

This requirement, of deciding at the
time of dumping the BDD, which nodes
will need to be mapped, instead of
at the time of loading the BDD,
does change the requirements of
the method `dd.bdd.BDD.load()`.

Nonetheless, explicit is better than implicit,
so explicitly specifying in the argument `roots`,
which nodes need to be mapped, aligns
with explicitly defining requirements.

This is a requirement by only `dd.bdd.BDD.load()`.
It is not a limitation of what information is
stored in the Pickle file.

If the mapping from all the dumped
nodes to all the loaded nodes is
found to be needed *after* dumping
the BDD, then this mapping can
be constructed by either:

- directly calling the (internal)
  `dd.bdd.BDD._load_pickle()`,
  and using item 0 of the tuple
  that it returns, or

- using the same code that is currently in
  `dd.bdd.BDD._load_pickle()`
  (namely the variable `umap` is this mapping),

and applying to the nodes of interest
the mapping that is represented by
the dictionary `umap`.


Part of the changes in this commit is:

- API: store roots in `dd.bdd.BDD._dump_bdd()`

  This commit changes the contents of
  the Pickle file where the BDD is dumped.
  The roots are in a `list`.

These two changes (in what is stored in the
Pickle file, and what is returned from the
method `BDD.load()`) form a single commit
in order to ensure that attempting to load
Pickle files dumped with `dd` versions earlier
to this change will result in a `KeyError`.

Other changes:

- TST: update tests relevant to the above changes