BUG: correct docstring of `dd._abc.BDD.load()`
Both the description (`@return` annotation) and
the type (`@rtype` annotation) where in error.

The description was incorrect because it described
the behavior of the method `BDD.load()` for Pickle
files (which used to be to return a `dict` that
was a mapping from dumped to loaded nodes),
but not for JSON files (for which a `list` of
root nodes is returned).

The API is currently homogeneous (after the
recent API change for Pickle files),
so the docstring is now simpler.

The docstring *will* change in an upcoming
extension of the API to support `dict`
containers for roots of BDDs when dumping
to files. That change will be for
both JSON and Pickle files.

The upcoming API change only adds supported
representations for the roots,
in a backwards compatible way (to the current API).