ENH: support roots as `dict` in Pickle interface
- API: Previously, the returned `dict` was a mapping from
  dumped to all loaded BDD nodes, whereas now it is
  a mapping from names (as strings) to loaded roots
  (BDD node references).

The change of meaning for the returned value
is not an issue for files dumped from earlier `dd`
versions, because attempting to load a BDD from
a Pickle file that was dumped with an earlier version
of `dd` would raise an error,
specifically when attempting to access the key `'root'`
of the unpickled `dict` (i.e., `d['roots']`).

This API change *is* an issue for files dumped
with this version of `dd`, in case the callee of
the method `BDD.load()` assumes the old interface
of the methods:

- `dd.bdd.BDD.load()`
- `dd.autoref.BDD.load()`.

There is no way to detect this case.

- REF: change `dd.cudd.BDD.dump()` to `def` (was `cpdef`)
  to use closures (for the variable `cache`).