API: build `dddmp` parser upon first call to `parse()`
instead of when instantiating the class `Parser`.

Initializing the attribute `Parser.parser` to `None` in
the function `Parser.__init__()`, and building
the parser upon first call to the method `Parser.parse()`
is similar to how the attribute `dd.bdd.BDD._parser`
was used in commit:
`4977e2438e90c0c0f4f02786cd093f18f3e4a0a0`

In more detail, the attribute `BDD._parser` was
initialized to `None` within the method `BDD.__init__()`,
and a parser was initialized upon first call to
the method `BDD.add_expr()`, and assigned to `BDD._parser`.