TYP: specify type hints for instance variables of `dd.bdd.BDD`
for readability, following the recommendations of [PEP 526](
    https://peps.python.org/pep-0526/#class-and-instance-variable-annotations).

Define and use `_Nat` as type alias for `int`,
to specify the set of natural numbers,
defined as `Nat` in the TLA+ module `Naturals`,
in the book "Specifying Systems".

In the future, annotations of Python code with
TLA+ expressions could be used for reasoning
about the algorithms described by the Python code,
for example via theorem proving.