API: measure memory via `os.sysconf()`.
Also, make the relevant check run when the relevant
information about system memory is available.
When `os.sysconf()` cannot read `SC_PAGE_SIZE` or `SC_PHYS_PAGES`,
the test is skipped with a message.

Previously, importing the package `psutil` was required.

The memory measurement is used to check that the initial
memory estimate given upon instantiation of the classes:
- `dd.cudd.BDD`
- `dd.cudd_zdd.BDD`
is within the memory that is available in the system.

- REL: rm `psutil` from `install_requires`