_API: turn `dd._abc` classes to `typing` protocols
The module `tests/inspect_cython_signatures.py`
already checks that the module `dd.cudd` implements
the module `dd._abc`.

`BDD` and `Function` in the module `dd.autoref`
inherit from the classes they implement, which are
specified in the module `dd._abc`.

The classes in `dd.cudd` do not inherit from
`dd._abc` classes because this is not supported in Cython.
Also, abstract base classes for `cdef` classes
are not supported by Cython.

This is the reason why `inspect_cython_signatures.py`
exists. Nonetheless, now that `typing.Protocol` is
available, and `dd` requires a Python 3 version that
includes `typing.Protocol`, it can be used.