BUG: implement `BDD.declare()` only in concrete classes
`dd._abc.BDD` is a protocol that specifies the
interface of the class `BDD`. This interface
definition does not include method `add_var()`.

`BDD.add_var()` is implemented in:
- `dd.bdd.BDD`
- `dd.autoref.BDD`

(`add_var()` exists altogether because
I introduced `declare()` later than
when `add_var()` was introduced.)

This error was found by `pytype`.