BUG: avoid duplication of item in `LD_LIBRARY_PATH`
The duplication occurred not in the value of the
environment variable `LD_LIBRARY_PATH` within the
current GHA step, but in the value of `LD_LIBRARY_PATH`
that got defined in `GITHUB_ENV` (for subsequent GHA steps).

The path added to `LD_LIBRARY_PATH` is now defined
at one place, instead of two.