API: rm circular dependency of `dd` examples on `omega`
Thus enabling CI-testing of `dd` examples.

Otherwise, backwards-incompatible changes in `dd`
would require CI to pass, before being released,
which would depend on a version of `omega` that
cannot yet have been updated to the yet-unreleased `dd`.

Even though the dependence was on the module
`omega.logic.syntax`, which is independent of
the package `dd` (transitively via `import` statements),
nonetheless this change avoids a circular dependency
of the `dd` repository (its directory `examples/`) and
the repository of `omega`.