MNT: assert that nodes are not `None`
nor `0` (nodes are nonzero).
It is for `dd.bdd.BDD._succ`:

```python
self._succ: dict[_Node, _Fork]
_Fork: _ty.TypeAlias = tuple[
    _Level,
    _Ref | None,
    _Node | None]
```

`None` is possible in a `_Fork` tuple.
In the locations where this commit adds
assertions, it is provable that:

```tlapy
THEOREM NotNone ==
    /\ u is not None
    /\ w is not None
```

However, this theorem is not deducible by
the type hints alone. So static analysis
based on type hints raises errors, which
motivated introducing these assertions.