BUG: ensure nodes are not `None` in `BDD._swap_cofactor()`
This error was found by `mypy`.

The method `dd.bdd.BDD._swap_cofactor()` is called
only from the method `dd.bdd.BDD.swap()`.
These calls were correct, in that `v` and `w` are
provably not `None` there.

In general though, `_swap_cofactor()` did not
ensure that it would not return
a `tuple` containing `None`. This is why this
commit is described as `BUG`.